requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
az-iranian-bank-gateways==1.9.0
billiard==4.1.0
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==4.2.5
django-admin-notification==0.2
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-colorfield==0.10.1
django-etc==1.4.0
django-hitcount==1.3.5
django-jalali==6.0.1
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-redis==5.3.0
django-render-partial==0.4
django-social-share==2.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
fonttools==4.42.1
html5lib==1.1
idna==3.4
ipcalc==1.99.0
isodate==0.6.1
jdatetime==4.1.1
kavenegar==1.1.2
kombu==5.3.2
lxml==4.9.3
Pillow==10.0.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
pycparser==2.21
pycryptodome==3.19.0
pydyf==0.7.0
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
zeep==4.2.1
zopfli==0.2.3
